requirements.txt
ultralytics
streamlit
opencv-python-headless
Pillow
numpy

